test for user.js

+++ packages/app/controllers/users.js
@@ -146,9 +146,25 @@ async function generateKeyController(req, res, next) {
 
     const { userId } = req.userData;
     const user = await userService.getUser(userId);
+    if (!user) {
+      return res.status(404).json({
+        statusCode: 404,
+        error: STATUS_CODES[404],
+        message: "User not found",
+      });
+    }
+
     const subscription = await subscriptionService.getSubscription(
       user.subscription_id
     );
+    if (!subscription) {
+      return res.status(206).json({
+        statusCode: 206,
+        error: STATUS_CODES[206],
+        message: "User Data not found",
+      });
+    }
+
     if (user.keys.length >= subscription.key_limit) {
       return res.status(403).json({
         message: "Limit reached. Consider upgrading your plan",
@@ -185,6 +201,14 @@ async function destroyKeyController(req, res, next) {
 
     const { userId } = req.userData;
     const user = await userService.getUser(userId);
+    if (!user) {
+      return res.status(404).json({
+        statusCode: 404,
+        error: STATUS_CODES[404],
+        message: "User not found",
+      });
+    }
+
     const destroyedKey = await userService.destroyUserKey(value.keyId, user);
     if (!destroyedKey) {
       return res.status(404).json({
@@ -221,6 +245,14 @@ async function updatePasswordController(req, res, next) {
     const { currPassword, newPassword } = value;
     const { email } = req.userData;
     const user = await userService.getUserByEmail(email);
+    if (!user) {
+      return res.status(404).json({
+        statusCode: 404,
+        error: STATUS_CODES[404],
+        message: "User not found",
+      });
+    }
+
     const matchPassword = await user.matchPassword(currPassword);
     if (!matchPassword) {
       return res.status(400).json({
